Remove "Col" + Utility Class, Logger Example
Col was the last utility left in the class, and since it is no longer used, both have been removed. The file is renamed to "UtilityExtensions", since that is now the only class present in it.
